fix login screen so it shows loading instead of login page if user is logged in
