+++ client/scrapescholar_client/app/page.tsx
@@ -1,22 +1,53 @@
 "use client";
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import Head from 'next/head';
 import SearchView from "./views/SearchView";
 import Login from './views/Login';
+import apiCalls from './api/apiCalls';
 export default function Home() {
   const [loggedIn, setLoggedIn] = useState<boolean>(false);
+  const [token, setToken] = useState('');
 
+  const [loadingLogin, setLoadingLogin]=useState(false);
+  const {getCookie } = apiCalls();
+  useEffect(() =>{
+    const fetchCookie = async () =>
+    {
+      try{
+    const cookie = await getCookie();
+    if(cookie.detail=="Cookie not found" || cookie.detail=="Invalid token"){
+        console.log("no cookie")
+        setLoggedIn(false);
+    }else{
+        setToken(cookie.detail)
+        setLoggedIn(true)
+    }}
+    catch(error){
+      console.error("Error fetching cookie:", error);
+    }
+    finally{
+      setLoadingLogin(true);
+    }
+    }
+    fetchCookie();
+    
+},[])
   
 
   return (
     <div>
-
-      {loggedIn ? (<SearchView setLoggedIn={setLoggedIn} />
+        {!loadingLogin?<>Loading</>:<>
+          {loggedIn ? (<SearchView setLoggedIn={setLoggedIn} />
       ) : (
         <>
-          <Login setLoggedIn={setLoggedIn}/>
+          <Login setLoggedIn={setLoggedIn} setToken={setToken}/>
         </>
       )}
+        
+        
+        
+        </>}
+    
 
     </div>
   );
